insight_2ontology/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-cloud-storage>=2.16.0
google-cloud-pubsub>=2.21.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-experimental>=0.3.0
langchain-google-genai>=2.0.0
